-- sweet alert implementation in post section.

+++ script/controllers/postController.js
@@ -3,7 +3,7 @@
  */
 (function () {
 
-    webpanel.controller('postController', ['$scope', '$videoService', function ($scope, $videoService) {
+    webpanel.controller('postController', ['$scope', '$videoService','sweet', function ($scope, $videoService,sweet) {
         $scope.showList = function () {
             var data = $videoService.showAllVideos();
             data.then(function (success) {
@@ -15,6 +15,7 @@
                 }
             }, function (error) {
                 //
+                sweet.show('Error', 'Can\'t process your request :(', 'error');
                 console.log("PostController", error);
             })
         };
@@ -34,10 +35,10 @@
                     if(success.status === "success"){
                         $scope.videos[index].post_status = "APPROVED";
                     }else{
-                        alert("Can't process your request. Try again.")
+                        sweet.show('Error', 'Can\'t process your request :(', 'error');
                     }
                 },function(error){
-                    alert("Can't process your request. Try again.")
+                    sweet.show('Error', 'Can\'t process your request :(', 'error');
                 });
             }
 
@@ -54,27 +55,44 @@
                     if(success.status === "success"){
                         $scope.videos[index].post_status = "REJECTED";
                     }else{
-                        alert("Can't process your request. Try again.")
+                        sweet.show('Error', 'Can\'t process your request :(', 'error');
                     }
                 },function(error){
-                    alert("Can't process your request. Try again.")
+                    sweet.show('Error', 'Can\'t process your request :(', 'error');
                 });
             }
         };
 
         $scope.delete = function (index, video_id) {
-            var deletedVideo = $videoService.deleteVideo(video_id);
-            deletedVideo.then(function (success) {
-                if (success.status === "success") {
-                    $scope.videos.splice(index, 1);
-                } else {
-                    alert("Something wrong!!! Unable to delete the video.");
-                }
-            }, function (error) {
-                alert("Something wrong!!! Unable to delete the video.");
-            });
-            console.log($scope.videos[index]);
+            sweet.show({
+                title: 'Confirm',
+                text: 'Delete this Video?',
+                type: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#DD6B55',
+                confirmButtonText: 'Yes, delete it!',
+                closeOnConfirm: false,
+                closeOnCancel: false
+            }, function(isConfirm) {
+                if (isConfirm) {
+                    var deletedVideo = $videoService.deleteVideo(video_id);
+                    deletedVideo.then(function (success) {
+                        if (success.status === "success") {
+                            $scope.videos.splice(index, 1);
+                            sweet.show('Deleted!', 'The video file has been deleted.', 'success');
+                        } else {
+                            sweet.show('Error', 'Unable to delete the video.', 'error');
+                            console.log(success);
+                        }
+                    }, function (error) {
+                        sweet.show('Error', 'Unable to delete the video.', 'error');
+                    });
+                    console.log($scope.videos[index]);
 
+                }else{
+                    sweet.show('Cancelled', 'Your video file is safe :)', 'error');
+                }
+            });
         };
 
         // showAllVideos
